Рефакторит render-функцию
